requirements.txt
asgiref==3.7.2
sqlparse==0.4.4
gunicorn==21.2.0
psycopg2-binary==2.9.9  # Адаптер для работы с PostgreSQL
django-environ==0.11.2
djangorestframework==3.12.4
drf-yasg==1.20.0
django-filter==2.4.0
django-cors-headers==3.8.0